# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ariadne Metadata Management API
  version: 0.1.0
  description: >
    Phase 0 contract for Ariadne backend. Covers authentication, metadata CRUD,
    lineage management (with lineage_source), bulk import/export, semantic search,
    and AI assistant endpoints. All endpoints are prefixed with `/api/v1`.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /sources:
    get:
      tags: [sources]
      summary: List data sources
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - in: query
          name: type
          schema:
            $ref: '#/components/schemas/SourceType'
        - in: query
          name: is_active
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated sources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [sources]
      summary: Create data source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '409':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /sources/test-connection:
    post:
      tags: [sources]
      summary: Test connection for a source configuration (no persistence)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreateRequest'
      responses:
        '200':
          description: Connection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionTestResponse'
        default:
          $ref: '#/components/responses/Error'
  /sources/{source_id}/test:
    post:
      tags: [sources]
      summary: Test connection for an existing source
      parameters:
        - $ref: '#/components/parameters/source_id'
      responses:
        '200':
          description: Connection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionTestResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /sources/{source_id}/introspect/table:
    post:
      tags: [sources]
      summary: Introspect a table/collection/index for a source
      parameters:
        - $ref: '#/components/parameters/source_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [table_name]
              properties:
                table_name:
                  type: string
                schema_name:
                  type: string
                  description: Optional schema/namespace for SQL sources
      description: |
        Requires bearer auth. For Oracle, ensure source.connection_config includes host, port, service_name, username, password (or dsn).
      responses:
        '200':
          description: Introspection result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /sources/{source_id}:
    get:
      tags: [sources]
      summary: Get source by id
      parameters:
        - $ref: '#/components/parameters/source_id'
      responses:
        '200':
          description: Source detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [sources]
      summary: Update source
      parameters:
        - $ref: '#/components/parameters/source_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpdateRequest'
      responses:
        '200':
          description: Updated source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [sources]
      summary: Delete (soft-delete) source
      parameters:
        - $ref: '#/components/parameters/source_id'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
  /tables:
    get:
      tags: [tables]
      summary: List tables
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - in: query
          name: source_id
          schema:
            type: string
            format: uuid
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
        - in: query
          name: type
          schema:
            type: string
      responses:
        '200':
          description: Paginated tables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [tables]
      summary: Create table metadata (manual)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        default:
          $ref: '#/components/responses/Error'
  /tables/{table_id}:
    get:
      tags: [tables]
      summary: Get table detail
      parameters:
        - $ref: '#/components/parameters/table_id'
      responses:
        '200':
          description: Table detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [tables]
      summary: Update table metadata
      parameters:
        - $ref: '#/components/parameters/table_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [tables]
      summary: Delete table metadata
      parameters:
        - $ref: '#/components/parameters/table_id'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
  /tables/{table_id}/fields:
    get:
      tags: [fields]
      summary: List fields for a table
      parameters:
        - $ref: '#/components/parameters/table_id'
      responses:
        '200':
          description: Fields of the table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [fields]
      summary: Create field metadata for a table (table_id from path)
      parameters:
        - $ref: '#/components/parameters/table_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldCreateInTable'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        default:
          $ref: '#/components/responses/Error'
  /tables/{table_id}/fields/batch:
    post:
      tags: [fields]
      summary: Batch create fields for a table (table_id from path)
      parameters:
        - $ref: '#/components/parameters/table_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/FieldCreateInTable'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Field'
        default:
          $ref: '#/components/responses/Error'
...
  /lineage/table:
    post:
      tags: [lineage]
      summary: Create table-level lineage relationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableLineageCreateRequest'
      responses:
        '201':
          description: Created lineage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageRelationship'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /tables/{table_id}/tags:
    get:
      tags: [tags]
      summary: Get tags for a table
      parameters:
        - $ref: '#/components/parameters/table_id'
      responses:
        '200':
          description: Table tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [tags]
      summary: Add tags to table
      parameters:
        - $ref: '#/components/parameters/table_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableTagsAddRequest'
      responses:
        '201':
          description: Tags added
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /tables/{table_id}/tags/{tag_id}:
    delete:
      tags: [tags]
      summary: Remove tag from table
      parameters:
        - $ref: '#/components/parameters/table_id'
        - $ref: '#/components/parameters/tag_id'
      responses:
        '204':
          description: Tag removed
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    size:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    source_id:
      in: path
      name: source_id
      required: true
      schema:
        type: string
        format: uuid
    table_id:
      in: path
      name: table_id
      required: true
      schema:
        type: string
        format: uuid
...
    tag_id:
      in: path
      name: tag_id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SourceType:
      type: string
      enum: [oracle, mongodb, elasticsearch, other]
    Source:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        description:
          type: string
        connection_config:
          type: object
          description: Redacted connection information
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SourceCreateRequest:
      type: object
      required: [name, type, connection_config]
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        description:
          type: string
        connection_config:
          type: object
          description: |
            Required fields by source type:
              - oracle: host, port (default 1521), service_name, username, password (or dsn)
              - mongodb: uri (e.g. mongodb://user:pass@host:port/db), database (optional)
              - elasticsearch: hosts (string or array), username/password (optional)
          additionalProperties: true
    SourceUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/SourceCreateRequest'
      properties:
        is_active:
          type: boolean
    ConnectionTestResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        latency_ms:
          type: integer
...
    SourceListResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Source'
    Table:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        name:
          type: string
        schema_name:
          type: string
          description: Logical schema or database namespace
        qualified_name:
          type: string
          description: Fully qualified name, e.g., schema.table
        type:
          type: string
          example: table
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        row_count:
          type: integer
        field_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TableDetail:
      allOf:
        - $ref: '#/components/schemas/Table'
      properties:
        fields:
          type: array
          items:
            $ref: '#/components/schemas/Field'
    TableCreateRequest:
      type: object
      required: [name, source_id]
      properties:
        source_id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    TableUpdateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    TableListResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Table'
    Field:
      type: object
      properties:
        id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        name:
          type: string
        data_type:
          type: string
        description:
          type: string
        is_nullable:
          type: boolean
        is_primary_key:
          type: boolean
        is_foreign_key:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FieldListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Field'
...
    LineageSource:
      type: string
      enum: [inferred, manual, approved]
    LineageRelationship:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [table, field]
        source_node_id:
          type: string
          format: uuid
        target_node_id:
          type: string
          format: uuid
        transformation_type:
          type: string
          example: ETL
        transformation_logic:
          type: string
        lineage_source:
          $ref: '#/components/schemas/LineageSource'
        status:
          type: string
          enum: [pending, approved, rejected]
          default: pending
        approved_by:
          type: string
          format: email
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        confidence:
          type: number
          format: float
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    TableLineageCreateRequest:
      type: object
      required: [source_table_id, target_table_id, lineage_source]
      properties:
        source_table_id:
          type: string
          format: uuid
        target_table_id:
          type: string
          format: uuid
        transformation_type:
          type: string
        transformation_logic:
          type: string
        lineage_source:
          $ref: '#/components/schemas/LineageSource'
        confidence:
          type: number
          minimum: 0
          maximum: 1
        metadata:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        level:
          type: integer
          minimum: 1
          maximum: 3
        path:
          type: string
          description: Full path like 'F10-公司概况-公司资料'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TagListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
...
    TableTagsAddRequest:
      type: object
      required: [tag_ids]
      properties:
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
    FieldCreateInTable:
      type: object
      required: [name, data_type]
      properties:
        name:
          type: string
        data_type:
          type: string
        description:
          type: string
        is_nullable:
          type: boolean
        is_primary_key:
          type: boolean
        is_foreign_key:
          type: boolean